Create ArduinoCode for LCD temperature display
This code initializes an LCD display using the LiquidCrystal library to show temperature readings in both Celsius and Fahrenheit. The temperature is measured using an LM35 sensor and is displayed on a 16x2 LCD module. The internal reference voltage (5V) is used for more accurate readings. The project also displays a startup message and updates the temperature every second